pyproject.toml
[tool.poetry]
name = "otel-play"
version = "0.1.0"
description = ""
authors = ["Sean-Miningah <seanpminingah@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
opentelemetry-instrumentation-django = "^0.46b0"
opentelemetry-sdk = "^1.25.0"
gunicorn = "^22.0.0"


[tool.poetry.group.dev.dependencies]
django-seed = "^0.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
